encoder refactoring
