+++ remoting/rico-remoting-common/src/main/java/dev/rico/internal/remoting/codec/encoders/ChangeAttributeMetadataCommandEncoder.java
@@ -40,7 +40,7 @@ public class ChangeAttributeMetadataCommandEncoder implements CommandTranscoder<
         jsonCommand.addProperty(ID, CHANGE_ATTRIBUTE_METADATA_COMMAND_ID);
         jsonCommand.addProperty(ATTRIBUTE_ID, command.getAttributeId());
         jsonCommand.addProperty(NAME, command.getMetadataName());
-        jsonCommand.add(VALUE, ValueEncoder.encodeValue(command.getValue()));
+        jsonCommand.add(VALUE, JsonUtils.encodeValue(command.getValue()));
         return jsonCommand;
     }
 
@@ -49,7 +49,7 @@ public class ChangeAttributeMetadataCommandEncoder implements CommandTranscoder<
         final ChangeAttributeMetadataCommand command = new ChangeAttributeMetadataCommand();
         command.setAttributeId(JsonUtils.getStringElement(jsonObject, ATTRIBUTE_ID));
         command.setMetadataName(JsonUtils.getStringElement(jsonObject, NAME));
-        command.setValue(ValueEncoder.decodeValue(jsonObject.get(VALUE)));
+        command.setValue(JsonUtils.decodeValue(jsonObject.get(VALUE)));
         return command;
     }
 }
